Splitting new model into different files and getting rid of more dead code
